automatically backed up by learn

+++ square_array.rb
@@ -1,6 +1,6 @@
 def square_array(array)
   array.each do {
-    |x|, x*x
+    |x.toInt|, x*x
   }
   end
   return array
